qualidade-software-seguranca-sistemas-inteligentes/MVP/backend/requirements.txt
Flask
Flask-Cors
waitress
python-multipart
scikit-learn
joblib
pandas
pdf2image
Pillow
pytesseract
scipy
pytest
